Upgrade to Rust 2024 edition (#5)
* Upgrade to Rust 2024

- Upgrade project to Rust 2024 edition
- Update dependencies: bitflags 2.6, bstr 1.11, criterion 0.5, libc 0.2
- Remove array-init dependency in favor of std::array::from_fn
- Fix clippy warnings and satisfy Rust 2024 unsafe requirements
- Add regression tests for jhash bug

* ci: update github actions and use cargo-llvm-cov for code coverage

* cargo fmt

* ci: trigger push only on main branch and v* tags

* fix: resolve new clippy warnings in FFI package

+++ benches/cqdb_bench.rs
@@ -1,7 +1,7 @@
 use std::{ffi::CString, fs, io::BufWriter};
 
-use cqdb::{CQDBWriter, CQDB};
-use criterion::{criterion_group, criterion_main, Criterion};
+use cqdb::{CQDB, CQDBWriter};
+use criterion::{Criterion, criterion_group, criterion_main};
 
 fn criterion_benchmark(c: &mut Criterion) {
     let mut group = c.benchmark_group("reader");
@@ -50,7 +50,7 @@ fn criterion_benchmark(c: &mut Criterion) {
         let db = CQDB::new(&buf).unwrap();
         b.iter(|| {
             for i in 0..db.num() {
-                let _value = db.to_str(i as u32).unwrap();
+                let _value = db.to_str(i).unwrap();
             }
         })
     });
